Производит рефакторинг кода.
