+++ src/main.js
@@ -8,13 +8,13 @@ import TripEventsComponent from './components/trip-events.js';
 import TripEventComponent from './components/trip-event.js';
 import TripEventEditComponent from './components/trip-event-edit.js';
 import NoEventsComponent from './components/no-events.js';
-import {renderComponent, replaceElement, addEscapeEvent} from './utils.js';
-import {FILTERS, MENU_ITEMS, RenderPosition} from './const.js';
+import {renderComponent, replaceComponent, RenderPosition} from './utils/render.js';
+import {addEscapeEvent} from './utils/common.js';
+import {FILTERS, MENU_ITEMS} from './const.js';
 import {generateEvents} from './mock/event.js';
 
 const EVENTS_COUNT = 10;
 
-let lastContainer = null;
 let lastEvent = null;
 let lastEditEvent = null;
 
@@ -29,10 +29,9 @@ const renderEvent = (container, event) => {
   const eventEditComponent = new TripEventEditComponent(event);
 
   const replaceEventToEdit = () => {
-    replaceElement(lastContainer.getElement(), lastEvent.getElement(), lastEditEvent.getElement());
+    replaceComponent(lastEvent, lastEditEvent);
     document.removeEventListener(`keydown`, documentEscPressHandler);
 
-    lastContainer = null;
     lastEvent = null;
     lastEditEvent = null;
   };
@@ -48,9 +47,8 @@ const renderEvent = (container, event) => {
       replaceEventToEdit();
     }
 
-    replaceElement(container.getElement(), eventEditComponent.getElement(), eventComponent.getElement());
+    replaceComponent(eventEditComponent, eventComponent);
 
-    lastContainer = container;
     lastEvent = eventComponent;
     lastEditEvent = eventEditComponent;
 
@@ -64,33 +62,33 @@ const renderEvent = (container, event) => {
   eventComponent.setRollupButtonHandler(eventEditHandler);
   eventEditComponent.setSubmitFormHandler(eventSubmitHadler);
 
-  renderComponent(container.getElement(), eventComponent.getElement());
+  renderComponent(container.getElement(), eventComponent);
 };
 
 const tripControls = document.querySelector(`.trip-main__trip-controls`);
 const tripInfo = document.querySelector(`.trip-main__trip-info`);
 
-renderComponent(tripControls, new MenuComponent(MENU_ITEMS).getElement(), RenderPosition.AFTERBEGIN);
-renderComponent(tripControls, new FilterComponent(FILTERS).getElement());
+renderComponent(tripControls, new MenuComponent(MENU_ITEMS), RenderPosition.AFTERBEGIN);
+renderComponent(tripControls, new FilterComponent(FILTERS));
 
 const tripEvents = document.querySelector(`.trip-events`);
 
 if (events.length === 0) {
-  renderComponent(tripEvents, new NoEventsComponent().getElement());
+  renderComponent(tripEvents, new NoEventsComponent());
 } else {
-  renderComponent(tripInfo, new TripInfoComponent(events).getElement(), RenderPosition.AFTERBEGIN);
+  renderComponent(tripInfo, new TripInfoComponent(events), RenderPosition.AFTERBEGIN);
 
-  renderComponent(tripEvents, new TripSortComponent().getElement());
+  renderComponent(tripEvents, new TripSortComponent());
 
   const tripDaysComponent = new TripDaysComponent();
-  renderComponent(tripEvents, tripDaysComponent.getElement());
+  renderComponent(tripEvents, tripDaysComponent);
 
   dates.forEach((date, index) => {
     const tripDayComponent = new TripDayComponent(date, index + 1);
-    renderComponent(tripDaysComponent.getElement(), tripDayComponent.getElement());
+    renderComponent(tripDaysComponent.getElement(), tripDayComponent);
 
     const tripEventsComponent = new TripEventsComponent();
-    renderComponent(tripDayComponent.getElement(), tripEventsComponent.getElement());
+    renderComponent(tripDayComponent.getElement(), tripEventsComponent);
 
     events
       .filter((event) => new Date(event.startDate).toDateString() === date)
