+++ src/utils/render.js
@@ -1,0 +1,30 @@
+export const RenderPosition = {
+  AFTERBEGIN: `afterbegin`,
+  BEFOREEND: `beforeend`
+};
+
+export const createElement = (template) => {
+  const element = document.createElement(`div`);
+
+  element.innerHTML = template;
+
+  return element.firstElementChild;
+};
+
+export const renderComponent = (container, component, position = RenderPosition.BEFOREEND) => {
+  switch (position) {
+    case RenderPosition.AFTERBEGIN:
+      container.prepend(component.getElement());
+      break;
+    case RenderPosition.BEFOREEND:
+      container.append(component.getElement());
+  }
+};
+
+export const replaceComponent = (newComponent, oldComponent) => {
+  const parentElement = oldComponent.getElement().parentElement;
+  const newElement = newComponent.getElement();
+  const oldElement = oldComponent.getElement();
+
+  parentElement.replaceChild(newElement, oldElement);
+};
